+++ src/components/DepositWidget/Form.tsx
@@ -8,7 +8,7 @@ import { TokenBalanceDetails } from 'types'
 import { formatAmountFull, parseAmount, abbreviateString } from 'utils'
 
 // Components
-import { CardDrawer } from 'components/layout/SwapLayout/Card'
+import { CardDrawer } from 'components/layout/LegacyTradeLayout/Card'
 import { Spinner } from 'components/common/Spinner'
 
 // DepositWidget: subcomponents
