+++ src/components/DepositWidget/index.tsx
@@ -12,7 +12,7 @@ import { TokenBalanceDetails } from 'types'
 // Components
 import ErrorMsg from 'components/ErrorMsg'
 import FilterTools from 'components/FilterTools'
-import { CardTable, CardWidgetWrapper } from 'components/layout/SwapLayout/Card'
+import { CardTable, CardWidgetWrapper } from 'components/layout/LegacyTradeLayout/Card'
 import { StandaloneCardWrapper } from 'components/layout'
 
 // DepositWidget: subcomponents
