+++ src/components/TradeWidget/TradeWidget.styled.ts
@@ -1,5 +1,5 @@
 import styled from 'styled-components'
-import Widget from 'components/layout/SwapLayout/Widget'
+import Widget from 'components/layout/LegacyTradeLayout/Widget'
 import { Wrapper } from 'components/ConnectWalletBanner'
 import { OrdersWrapper } from 'components/OrdersWidget/OrdersWidget.styled'
 import { WalletDrawerInnerWrapper } from 'components/DepositWidget/Form.styled'
