+++ src/components/layout/LegacyTradeLayout/Card/Card.tsx
