+++ src/components/layout/LegacyTradeLayout/Card/index.tsx
@@ -1,0 +1,1 @@
+export * from 'components/layout/LegacyTradeLayout/Card/Card'
