+++ src/components/layout/LegacyTradeLayout/Footer.tsx
