+++ src/components/layout/LegacyTradeLayout/Header/Header.styled.ts
