+++ src/components/layout/LegacyTradeLayout/Header/Navigation.styled.ts
