+++ src/components/layout/LegacyTradeLayout/Header/useOpenCloseNav.ts
