+++ src/components/layout/LegacyTradeLayout/LegacyTradeLayout.stories.tsx
@@ -3,20 +3,20 @@ import React from 'react'
 // also exported from '@storybook/react' if you can deal with breaking changes in 6.1
 import { Meta } from '@storybook/react/types-6-0'
 
-import { SwapLayout } from 'components/layout'
+import { LegacyTradeLayout } from 'components/layout'
 import { RouterDecorator } from 'storybook/RouterDecorator'
 import { LoremIpsum } from 'storybook/LoremIpsum'
 
 export default {
-  title: 'Layout/SwapLayout',
-  component: SwapLayout,
+  title: 'Layout/LegacyTradeLayout',
+  component: LegacyTradeLayout,
   decorators: [RouterDecorator],
 } as Meta
 
 export const Normal: React.FC = () => (
-  <SwapLayout>
+  <LegacyTradeLayout>
     <LoremIpsum />
-  </SwapLayout>
+  </LegacyTradeLayout>
 )
 
-export const ShortPage: React.FC = () => <SwapLayout>This is a really short page...</SwapLayout>
+export const ShortPage: React.FC = () => <LegacyTradeLayout>This is a really short page...</LegacyTradeLayout>
