+++ src/components/layout/LegacyTradeLayout/Widget.tsx
