--- src/components/layout/SwapLayout/Card/index.tsx
@@ -1,1 +1,0 @@
-export * from 'components/layout/SwapLayout/Card/Card'
